Se cambio forma de que los Helper reciben los datos
- Se agrego campo identification a la tabla Customers
- Los campos de precio y total pasaron a ser float
- Se modifico metodo de insert para que retorne el id generado
- Se unieron las pantallas de agregar cliente y agregar producto a las
de las listas

+++ src/com/diso/koala/db/Product.java
@@ -4,7 +4,7 @@ public class Product {
 
     private int id;
     private String name;
-    private int price;
+    private float price;
     private String barcode;
     private String description;
 
@@ -22,7 +22,7 @@ public class Product {
         this.name = name;
     }
 
-    public void setPrice(int price){
+    public void setPrice(float price){
         this.price = price;
     }
 
@@ -44,7 +44,7 @@ public class Product {
         return name;
     }
 
-    public int getPrice(){
+    public float getPrice(){
         return price;
     }
